requirements.txt
streamlit
openai>=1.0
pytesseract
Pillow
langchain
langchain-community